requirements.txt
langchain
langchain-community
langchain-groq
langsmith
python-dotenv